코멘트 중간점검 후 수정
